Delete archive (due to a fix in ID generation)
Previously the hash at the end of each link ID was not generated from
the source URL, but the link ID without the hash. The hash was therefore
redundant. Since the fix, the link IDs will be different, so this commit
deletes all entries in the archive. The subsequent commit will
re-archive links from the knowledge base.